+++ package.json
@@ -4,10 +4,12 @@
   "private": true,
   "scripts": {
     "build": "webpack --mode=production",
-    "serve": "webpack-dev-server"
+    "serve": "webpack-dev-server",
+    "lint": "eslint -c package.json src"
   },
   "dependencies": {
     "aframe": "^1.0.3",
+    "aframe-extras": "^6.1.0",
     "core-js": "^3.4.4",
     "vue": "^2.6.10",
     "vuex": "^3.1.2"
@@ -41,7 +43,8 @@
       "quotes": [
         "error",
         "double"
-      ]
+      ],
+      "no-console": "off"
     },
     "parserOptions": {
       "parser": "babel-eslint"
