+++ src/components/index.js
@@ -1,0 +1,8 @@
+// Afarame Components
+import "aframe"
+import "aframe-extras"
+
+// Afarame Components
+import "./appearance.js"
+import "./below-to-up.js"
+import "./player.js"
