+++ src/main.js
@@ -1,8 +1,6 @@
 import Vue from "vue"
-import App from "./App.vue"
-import AframeComponentStore from "./AframeComponentStore.js"
-import Aframe from "aframe"
-import Store from "./store/Store.js"
+import App from "./vue-components/App.vue"
+import store from "./store/Store.js"
 
 Vue.config.productionTip = false
 Vue.config.ignoredElements = [
@@ -13,14 +11,9 @@ Vue.config.ignoredElements = [
   "a-text",
 ]
 
-// event listenerの登録
-Object.keys(AframeComponentStore).forEach(function(component) {
-  Aframe.registerComponent(component, AframeComponentStore[component])
-})
-
 window.addEventListener("DOMContentLoaded", () => {
   new Vue({
-    store: Store,
+    store: store,
     render: h => h(App),
   }).$mount("#app")
 })
