+++ src/vue-components/AframeApp.vue
