+++ src/vue-components/Ghost.vue
