+++ src/vue-components/StraightRoad.vue
