+++ webpack.config.js
@@ -6,6 +6,7 @@ const CopyWebpackPlugin = require("copy-webpack-plugin")
 module.exports = {
   entry: {
     index: path.join(__dirname, "src", "main.js"),
+    "ghost-object": path.join(__dirname, "src", "components", "index.js")
   },
   output: {
     filename: "assets/js/[name]-[chunkhash].js"
